🐛 Fix Hardware Health card losing data on navigation
The Hardware Health card used raw useState/useEffect with no persistence,
causing data loss when navigating between dashboards. Converted to the
useCache system with IndexedDB persistence (stale-while-revalidate pattern)
so cached data survives navigation and shows instantly on return.

- Created useCachedHardwareHealth hook in useCachedData.ts
- Replaced 8 useState calls + 80-line fetch useEffect with single hook
- Added proper types, demo data, and parallel fetch for both endpoints
- Card now uses same caching pattern as all other dashboard cards
